+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Fixed some performance issues for users who follow a large number of accounts.
+- Tapping on a tab bar icon can let you scroll to top.
 
 ## [0.1 (95)] - 2023-11-27Z